Extract table_c2 callbacks
